Refactor per PR comments
