+++ src/api/operations/household.fragments.graphql
@@ -11,7 +11,19 @@ fragment HouseholdClientFields on HouseholdClient {
   id
   relationshipToHoH
   client {
-    ...HouseholdClientFieldsClient
+    id
+    ...ClientName
+    ...ClientIdentificationFields
+    veteranStatus
+    access {
+      ...ClientAccessFields
+    }
+    externalIds {
+      ...ClientIdentifierFields
+    }
+    alerts {
+      ...ClientAlertFields
+    }
   }
   enrollment {
     id
@@ -23,22 +35,6 @@ fragment HouseholdClientFields on HouseholdClient {
       id
       name
     }
-  }
-}
-
-fragment HouseholdClientFieldsClient on Client {
-  id
-  ...ClientName
-  ...ClientIdentificationFields
-  veteranStatus
-  access {
-    ...ClientAccessFields
-  }
-  externalIds {
-    ...ClientIdentifierFields
-  }
-  alerts {
-    ...ClientAlertFields
   }
 }
