+++ src/modules/enrollment/components/dashboardPages/HouseholdPage.tsx
@@ -5,7 +5,6 @@ import { useEnrollmentDashboardContext } from '@/components/pages/EnrollmentDash
 import NotFound from '@/components/pages/NotFound';
 import useSafeParams from '@/hooks/useSafeParams';
 import HouseholdMemberTable from '@/modules/household/components/HouseholdMemberTable';
-import { useHouseholdMembers } from '@/modules/household/hooks/useHouseholdMembers';
 
 const HouseholdPage = () => {
   const { enrollment } = useEnrollmentDashboardContext();
@@ -13,10 +12,6 @@ const HouseholdPage = () => {
     enrollmentId: string;
     clientId: string;
   };
-
-  const [householdMembers, { loading: householdMembersLoading, error }] =
-    useHouseholdMembers(enrollmentId);
-  if (error) throw error;
 
   if (!enrollment) return <NotFound />;
 
@@ -30,12 +25,7 @@ const HouseholdPage = () => {
         </CommonLabeledTextBlock>
       }
     >
-      <HouseholdMemberTable
-        householdMembers={householdMembers ? householdMembers : []}
-        householdMembersLoading={householdMembersLoading}
-        clientId={clientId}
-        enrollmentId={enrollmentId}
-      />
+      <HouseholdMemberTable clientId={clientId} enrollmentId={enrollmentId} />
     </TitleCard>
   );
 };
